push props down and remove handling unknown reading types
