+++ packages/webapp/public/locales/en/translation.json
@@ -1515,6 +1515,7 @@
     "HARDWARE_VERSION": "Hardware version",
     "NO_DATA_FOUND": "No sensor readings found",
     "NO_DATA": "(no data)",
+    "LAST_UPDATED": "Last updated {{latestReadingUpdate}}",
     "DETAIL": {
       "RETIRE": "Retire",
       "EDIT": "Edit",
@@ -1546,26 +1547,18 @@
     },
     "TEMPERATURE_READINGS_OF_SENSOR": {
       "TITLE": "Soil temperature",
-      "SUBTITLE": "Today’s ambient high and low temperature: {{high}}{{tempUnit}} / {{low}}{{tempUnit}}",
-      "Y_AXIS_LABEL": "Temperature in {{tempUnit}}",
+      "SUBTITLE": "Today’s ambient high and low temperature: {{high}}{{units}} / {{low}}{{units}}",
+      "Y_AXIS_LABEL": "Temperature in {{units}}",
       "WEATHER_STATION": "Weather station: {{weatherStationLocation}}",
-      "NO_DATA": "(no data)",
-      "AMBIENT_TEMPERATURE_FOR": "Ambient temperature for",
-      "LAST_UPDATED_TEMPERATURE_READINGS": "Last updated {{latestReadingUpdate}}"
+      "AMBIENT_TEMPERATURE_FOR": "Ambient temperature for"
     },
     "SOIL_WATER_POTENTIAL_READINGS_OF_SENSOR": {
       "TITLE": "Soil water potential",
-      "SUBTITLE": "MISSING",
-      "Y_AXIS_LABEL": "Soil water potential in {{soilWaterPotentialUnit}}",
-      "NO_DATA": "(no data)",
-      "LAST_UPDATED_SOIL_WATER_POTENTIAL_READINGS": "Last updated {{latestReadingUpdate}}"
+      "Y_AXIS_LABEL": "Soil water potential in {{units}}"
     },
     "SOIL_WATER_CONTENT_READINGS_OF_SENSOR": {
       "TITLE": "Soil water content",
-      "SUBTITLE": "MISSING",
-      "Y_AXIS_LABEL": "Soil water content in {{soilWaterContentUnit}}",
-      "NO_DATA": "(no data)",
-      "LAST_UPDATED_SOIL_WATER_CONTENT_READINGS": "Last updated {{latestReadingUpdate}}"
+      "Y_AXIS_LABEL": "Soil water content in {{units}}"
     },
     "READINGS_PREVIEW": {
       "TEMPERATURE": "Current soil temp",
