--- packages/webapp/src/containers/SensorReadings/styles.module.scss
@@ -1,25 +1,0 @@
-.titleWrapper{
-  margin: 24px 24px 0 12px;
-  display: flex;
-  justify-content: space-between;
-  align-items: baseline;
-  flex-wrap: wrap;
-}
-
-.title {
-  font-style: normal;
-  font-weight: normal;
-  color: var(--green);
-  font-size: 20px;
-  line-height: 24px;
-  text-transform: capitalize;
-  font-family: 'Open Sans', 'SansSerif', serif;
-}
-
-.emptyRect{
-  margin-top: 12px;
-  width: 100%;
-  height: 100px;
-  border: 1px solid var(--grey900);
-  border-style: dotted;
-}
